refactor: rename contact schema file from plural to singular
Rename contacts.ts to contact.ts for consistent singular naming convention
and update all import references across affected files.
