+++ src/components/pages/user-page/editable-display-name/display-name-editor/change-display-name.api.ts
@@ -4,7 +4,7 @@ import camelcaseKeys from 'camelcase-keys'
 import { revalidatePath } from 'next/cache'
 import snakecaseKeys from 'snakecase-keys'
 import { z } from 'zod'
-import { contactSchema } from '@/schemas/response/contacts'
+import { contactSchema } from '@/schemas/response/contact'
 import { ResultObject } from '@/types/api'
 import { fetchData } from '@/utils/api/fetch-data'
 import { getBearerToken } from '@/utils/cookie/bearer-token'
