+++ src/components/pages/user-page/editable-note/note-editor/change-note.api.ts
@@ -3,7 +3,7 @@
 import camelcaseKeys from 'camelcase-keys'
 import { revalidatePath } from 'next/cache'
 import { z } from 'zod'
-import { contactSchema } from '@/schemas/response/contacts'
+import { contactSchema } from '@/schemas/response/contact'
 import { ResultObject } from '@/types/api'
 import { fetchData } from '@/utils/api/fetch-data'
 import { getBearerToken } from '@/utils/cookie/bearer-token'
