+++ src/utils/api/get-contacts.ts
@@ -3,7 +3,7 @@ import camelcaseKeys from 'camelcase-keys'
 import { redirect } from 'next/navigation'
 import { cache } from 'react'
 import { z } from 'zod'
-import { contactSchema } from '@/schemas/response/contacts'
+import { contactSchema } from '@/schemas/response/contact'
 import { paginationDataSchema } from '@/schemas/response/pagination'
 import { fetchData } from '@/utils/api/fetch-data'
 import { getBearerToken } from '@/utils/cookie/bearer-token'
